разбиение таблицы
