+++ frontend/src/view/pages/timelineTasks/timeline/staticInfoPart/index.ts
@@ -1,0 +1,3 @@
+import TimelineTasksPartStaticInfoPart from './staticInfoPart';
+
+export { TimelineTasksPartStaticInfoPart };
