+++ frontend/src/view/pages/timelineTasks/timeline/tableTimeline/index.ts
@@ -1,0 +1,3 @@
+import TableTimeline from './tableTimeline';
+
+export { TableTimeline };
