+++ frontend/src/view/pages/timelineTasks/timeline/tableTimeline/styles.module.scss
@@ -1,0 +1,61 @@
+@import '../../../../theme/mixins';
+@import '../../../../theme/variables';
+
+.tableWrapper {
+  @include custom-scrollbars;
+  margin-left: 13rem;
+  
+  .table {
+    border-collapse: collapse;
+
+    .today {
+      background: lighten($accent-color, 20) !important;
+    }
+
+    .todayTd {
+      background: lighten($accent-color, 40) !important;
+    }
+  
+    .th {
+      border: 1px solid $grey-border-color;
+      padding: .5rem;
+      background: lighten($main-grey, 15);
+    }
+
+    .td {
+      border: 1px solid $grey-border-color !important;
+      background: white;
+
+      .content {
+
+        .label {
+          display: flex;
+          align-items: center;
+          padding: .5rem;
+
+          background: white;
+
+          border: 1px solid $grey-border-color;
+          border-radius: $main-radius;
+        }
+
+        height: 3rem;
+
+        display: flex;
+        align-items: center;
+        justify-content: space-between;
+        gap: .5rem;
+
+        [class~='p-button'] {
+          [class~='pi'] {
+            font-size: 1.3rem;
+          }
+        }
+      }
+    }
+
+    .filledTd {
+      background: wheat;
+    }
+  }
+}
